Schedule Request System
Small bugs found! Will be fixed in the next version.
Bugs: When user deletes a talk in the chosen screen he must re open schedule screen then chosen screen to see that it is deleted. And if user adds a topic to the already added ones he must also do the same to saw the change in the screen